+++ lte/gateway/c/oai/tasks/amf/amf_as.cpp
@@ -40,7 +40,7 @@ extern "C" {
 using magma5g::AsyncM5GAuthenticationServiceClient;
 
 using namespace magma;
-typedef uint32_t amf_ue_ngap_id_t;
+typedef uint64_t amf_ue_ngap_id_t;
 #define QUADLET 4
 #define AMF_GET_BYTE_ALIGNED_LENGTH(LENGTH)                                    \
   LENGTH += QUADLET - (LENGTH % QUADLET)
